del all js files and wrote all code for login, registration and get posts over again in only one file. No error and posts are in console
