--- js/api/fetch.js
@@ -1,8 +1,0 @@
-import { headers } from "./headers.js";
-
-export async function authFetch(url, options = {}) {
-	return fetch(url, {
-		...options,
-		headers: headers(Boolean(options.body)),
-	});
-}
