--- js/routes/home.js
@@ -1,5 +1,0 @@
-import { setAuthListener } from "../ui/listeners/auth.js";
-
-export async function homePage() {
-	setAuthListener()
-}
