--- js/ui/listeners/auth.js
@@ -1,5 +1,0 @@
-import { onAuth } from "../events/onAuth.js";
-
-export function setAuthListener() {
-	document.forms.auth.addEventListener("submit", onAuth);
-}
